Tree Assets size and animation change
I ran into issues with the animations due to the size of the original asset. I have to downsize everything and remake the sprites. They are now 1250x1500. I updated the animations for tree1 and tree2, now they have way more frames and resemble the intended video animation. I will be doing the same for the rest soon.